input_stream: convert offset and size to the off_t data type
size_t and long aren't 64 bit safe (i.e. files larger than 2 GB on a
32 bit OS).  Use off_t instead, which is a 64 bit integer if compiled
with large file support.

+++ src/decoder/mp3_plugin.c
@@ -673,7 +673,7 @@ static int decodeFirstFrame(mp3DecodeData * data,
 	 * Attempt to calulcate the length of the song from filesize
 	 */
 	{
-		size_t offset = data->inStream->offset;
+		off_t offset = data->inStream->offset;
 		mad_timer_t duration = data->frame.header.duration;
 		float frameTime = ((float)mad_timer_count(duration,
 		                   MAD_UNITS_MILLISECONDS)) / 1000;
